increase the margin of second box
With increasing the margin of the second box, the handling of margins gets even more clear. Since the margin was equal for both elements befor, the margin was simply collapsed and the distance stayed at 30px. Now that the margin of the second box was increased to 40px, the new distance between the objects will be 40px since the second box has the larger margin and therefore dominats the setting.