add property box-sizing to styling
This will change how the size is calculated. Since before the size of the element was calculated by adding up the size, the padding and the border, now the size provided in the width and height properties will be subtracted by padding and border and what remains will be used to render the content.