add basics of the box model (#14)
* add boilerplate to make boxes visible

* add boilerplate for first assignment

* add css file to header of index.html

* add two divs with different classes to index.html

* remove margin and padding globaly using css

* make font-size 40px page wide

* add background-color to boxes

* add padding to first box

the padding of 20px will enlarge the box, which previously was only ad large as the conteten, which filled it out. Now the padding will enlarge the box and add space equal to the set value (20px) around the content to the border of the box.

* add border to first box with color

By adding the border to the first box we make it visible. This comes on top of the previously defined padding and was colored to have a contrast between to the padding (which takes on the color of the element).

* add margin to first box

By adding the margin to an element, this element will be spaced from surrounding eleemnts. The messurement is there taken from the outside to the other outside of the borders of two neighboring elements.

* change values to different sizes for better differentiation

* add height and width to first box to make content to a fixed size

* add same styling to box to to showcase margin handling

By adding the same styling to the second box, the effect of margins collapsing on each other can be showcased, which will be displayed in the browser. The space betweeen the two boxes stays the same even after adding the additional margin to the second box.

* increase the margin of second box

With increasing the margin of the second box, the handling of margins gets even more clear. Since the margin was equal for both elements befor, the margin was simply collapsed and the distance stayed at 30px. Now that the margin of the second box was increased to 40px, the new distance between the objects will be 40px since the second box has the larger margin and therefore dominats the setting.

* add property box-sizing to styling

This will change how the size is calculated. Since before the size of the element was calculated by adding up the size, the padding and the border, now the size provided in the width and height properties will be subtracted by padding and border and what remains will be used to render the content.

* move result of first assignment to its own subdirectory

* add setup for second assignment

* remove external css to make working with it easier

* add block html for first example

* add html to block.html

* add border to p and ul type selectors

* add border to block class and li type selector

* add display properties to ul type selector and block class

* remove dots from unordered lists

* add exmple for inline usage